compare images by hash
